backend/requirements.txt
# CrewAI with Vertex AI for nurse scheduling
crewai>=0.86.0
crewai-tools>=0.17.0
google-cloud-aiplatform>=1.38.0
supabase>=2.0.0
python-dotenv>=1.0.0
litellm>=1.1.0
litellm[proxy]

# FastAPI for API endpoint
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Timefold solver
timefold>=1.0.0

# Langfuse tracing
langfuse>=2.0.0
openinference-instrumentation-crewai>=0.1.0

# Google ADK and Gemini
google-genai>=0.7.0
google-adk>=0.3.0

# File processing
openpyxl>=3.1.0
python-docx>=1.1.0
python-multipart>=0.0.9

# CopilotKit SDK for AG-UI protocol
copilotkit>=0.1.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
ag-ui-langgraph>=0.1.0

